calculate information out of the requested timestamps
So i wondered what the offsets were and implemented some arithemtic
to that effect.

After a while i noticed the offsets getting bigger and bigger because
stray packets came in and threw the code off. In order to make it more
robust against stray icmp packets i used the icmp sequence to match
the request with the reply.